workaround to get rid of illegal nested links
In HTML, it is illegal to nest <a> tags; this is unfortunate for us
because we will in general have nested links all the time.

The solution was to do some clever XSLT that flattens out the link
hierarchy, by moving the <a> tags to the text leaves of the tree; of
course, this would have very strange formatting, so we make <a> have no
formatting at all and format links explicitly using <span>.

The main edge case arises in the TOC, which I needed to handle using
JavaScript.